Fix EFAULT on some platforms.
The program header was marked as exec,
despite the fact that it contained
data that needed to be read by the kernel.
The write syscall would efault as the
kernel was unable to read the address.

On some versions of the kernel exec
is treated as read but on others it
only allows execution.

We now explicitly mark the section as read+exec.